continuing to move/refactor
