+++ commands/healthchecks/AlarmState.java
@@ -1,0 +1,6 @@
+package org.usfirst.frc4904.standard.commands.healthchecks;
+
+
+public enum AlarmState {
+	OK, INSUFFICIENT, ALARM;
+}
