+++ commands/healthchecks/BodyHealthLevelProvider.java
@@ -1,0 +1,5 @@
+package org.usfirst.frc4904.standard.commands.healthchecks;
+
+public class BodyHealthLevelProvider {
+
+}
