+++ commands/healthchecks/BodyPartsSongGenerator.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.standard.commands.safety;
+package org.usfirst.frc4904.standard.commands.healthchecks;
 
 public interface BodyPartsSongGenerator {
 	public BodyPartsSong generateSong(Body body);
