+++ commands/healthchecks/HealthProtectCommand.java
