+++ commands/healthchecks/InterruptCompressor.java
