--- temp-from-leif-health-checks/AlarmState.java
@@ -1,6 +1,0 @@
-package org.usfirst.frc4904.standard.commands.safety;
-
-
-public enum AlarmState {
-	OK, INSUFFICIENT, ALARM;
-}
