--- temp-from-leif-health-checks/BodyHealthLevelProvider.java
@@ -1,5 +1,0 @@
-package org.usfirst.frc4904.standard.commands.safety;
-
-public class BodyHealthLevelProvider {
-
-}
